+++ exjs.exs
@@ -3,7 +3,9 @@ defmodule ElixirScript.Config do
   def project() do
     [
       app: :app,
-      js_deps: []
+      js_deps: [
+        { :JQuery, from: "jquery" }
+      ]
     ]
   end
 end