+++ lib/elixir_script/translator/pattern_matching.ex
@@ -102,10 +102,10 @@ defmodule ElixirScript.Translator.PatternMatching do
             Translator.translate(x),
             Builder.call_expression(
               Builder.member_expression(
-                ref,
-                Builder.identifier(:get)
+                Builder.identifier(:Kernel),
+                Builder.identifier(:elem)
               ),
-              [Builder.literal(index)]
+              [ref, Builder.literal(index)]
             )
           )
 
@@ -316,10 +316,10 @@ defmodule ElixirScript.Translator.PatternMatching do
                 Builder.identifier(item),
                 Builder.call_expression(
                   Builder.member_expression(
-                    identifier_fn.(index),
-                    Builder.identifier(:get)
+                    Builder.identifier(:Kernel),
+                    Builder.identifier(:elem)
                   ),
-                  [Builder.literal(current_state.state_index)]
+                  [identifier_fn.(index), Builder.literal(current_state.state_index)]
                 )
               )
           declaration = Builder.variable_declaration([declarator], :let)
@@ -330,10 +330,10 @@ defmodule ElixirScript.Translator.PatternMatching do
             fn(new_index) ->
               Builder.call_expression(
                 Builder.member_expression(
-                  identifier_fn.(index),
-                  Builder.identifier(:get)
+                  Builder.identifier(:Kernel),
+                  Builder.identifier(:elem)
                 ),
-                [Builder.literal(current_state.state_index + new_index)]
+                [identifier_fn.(index), Builder.literal(current_state.state_index + new_index)]
               )
           end, nil)